+++ tests/utils/test_saliency_maps_gcn.py
@@ -88,9 +88,12 @@ def create_GCN_model_sparse(graph):
 def test_ig_saliency_map():
     graph = example_graph_1(feature_size=4)
     base_model, keras_model_gcn, generator, train_gen = create_GCN_model_dense(graph)
-    base_model_sp, keras_model_gcn_sp, generator_sp, train_gen_sp = create_GCN_model_sparse(
-        graph
-    )
+    (
+        base_model_sp,
+        keras_model_gcn_sp,
+        generator_sp,
+        train_gen_sp,
+    ) = create_GCN_model_sparse(graph)
 
     keras_model_gcn.compile(
         optimizer=Adam(lr=0.1), loss=categorical_crossentropy, weighted_metrics=["acc"]
@@ -156,9 +159,12 @@ def test_ig_saliency_map():
 def test_saliency_init_parameters():
     graph = example_graph_1(feature_size=4)
     base_model, keras_model_gcn, generator, train_gen = create_GCN_model_dense(graph)
-    base_model_sp, keras_model_gcn_sp, generator_sp, train_gen_sp = create_GCN_model_sparse(
-        graph
-    )
+    (
+        base_model_sp,
+        keras_model_gcn_sp,
+        generator_sp,
+        train_gen_sp,
+    ) = create_GCN_model_sparse(graph)
 
     keras_model_gcn.compile(
         optimizer=Adam(lr=0.1), loss=categorical_crossentropy, weighted_metrics=["acc"]